make sure uid is bigint

+++ db/migrate/20200803022845_create_users.rb
@@ -1,7 +1,7 @@
 class CreateUsers < ActiveRecord::Migration[6.0]
   def change
     create_table  :users do |t|
-      t.integer   :uid,         null: false
+      t.bigint    :uid,         null: false
       t.string    :cookie_key,  null: false
       t.jsonb     :json,        null: false,  default: '{}'
       t.timestamps
